refactor: Move modules under lib directory
Move module files from the project root to the `lib` directory.
