+++ lib/RPCError.js
