+++ lib/readJson.js
